requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.7
cffi==1.12.3
coverage==4.5.3
cryptography==2.7
cssselect==1.0.3
decorator==4.4.0
Django==2.2.3
django-cors-headers==3.1.1
django-filter==2.2.0
django-model-utils==3.2.0
djangorestframework==3.10.3
entrypoints==0.3
fabric==2.4.0
flake8==3.7.8
importlib-metadata==0.18
invoke==1.2.0
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
lxml==4.4.0
Markdown==3.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.0
paramiko==2.6.0
parso==0.5.1
pathlib2==2.3.4
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pluggy==0.12.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
PyNaCl==1.3.0
pyparsing==2.4.0
pytest==5.0.1
pytest-cov==2.7.1
pytest-django==3.5.1
pytest-flake8==1.0.4
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
zipp==0.5.2
